backend/requirements.txt
PyJWT==2.6.0
SQLAlchemy==1.4.45
